feat: 회원가입 및 로그인 post 기능 및 오류 추가
